refactor: 에브리타임 패키지 구조 1차 수정
