+++ backend/src/main/java/unischedule/everytime/service/EverytimeService.java
@@ -4,11 +4,11 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import unischedule.everytime.client.OpenAiClient;
+import unischedule.external.OpenAiClient;
 import unischedule.everytime.dto.TimetableDetailDto;
 import unischedule.everytime.dto.TimetableDto;
-import unischedule.everytime.dto.external.EverytimeTimetableRawResponseDto;
-import unischedule.everytime.client.EverytimeClient;
+import unischedule.external.dto.EverytimeTimetableRawResponseDto;
+import unischedule.external.EverytimeClient;
 import unischedule.everytime.mapper.EverytimeTimetableMapper;
 import unischedule.exception.ExternalApiException;
 import unischedule.exception.InvalidInputException;
